# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Game Collector App
# during dev, should point to your local machine
#host: localhost:10010
host: advance-engine-250015.appspot.com

x-google-endpoints:
- name: 'advance-engine-250015.appspot.com'
  allowCors: 'true'
# basePath prefixes all resource paths
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:

  /gamesystems:
    x-swagger-router-controller: gamesystem.controller

    get:
      description: get the game system list
      parameters:
        - name: name
          in: query
          type: string
          required: false
        - name: sort
          in: query
          type: string
          required: false
      operationId: getGameSystems
      responses: 
        "200":
          description: success
          schema: 
            $ref: "#/definitions/GetGameSystemListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    post:
      description: Add a new game system to the list
      parameters:
        - in: body
          name: gamesystem
          description: The game system to create
          schema:
            $ref: "#/definitions/GameSystem"
      operationId: createGameSystem
      responses: 
        "201":
          description: success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
  
  /gamesystems/{id}:
    x-swagger-router-controller: gamesystem.controller
    
    get: 
      description: Get a game system
      parameters:
        - name: id
          type: string
          in: path
          required: true
      operationId: getGameSystemById
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: Update a game system
      parameters:
        - name: id
          description: Game system id
          type: string
          in: path
          required: true
        - in: body
          name: gamesystem
          description: The game system to update
          schema:
            $ref: "#/definitions/GameSystem"
      operationId: updateGameSystem
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      description: Delete a game system
      parameters:
        - name: id
          description: Game system id
          type: string
          in: path 
          required: true
      operationId: deleteGameSystem
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /videogames:
    x-swagger-router-controller: game.controller
    get:
      description: get the game list
      parameters:
        - name: name
          in: query
          type: string
          required: false
        - name: developer
          in: query
          type: string
          required: false
        - name: gamesystem
          in: query
          type: string
          required: false
        - name: genre
          in: query
          type: string
          required: false
        - name: year
          in: query
          type: number
          required: false
        - name: sort
          in: query
          type: string
          required: false
        - name: fields
          in: query  
          type: string
          required: false
      operationId: getGames
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetVideoGameListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new game system to the list
      parameters:
        - in: body
          name: videogame
          description: The videogame to create.
          schema:
            $ref: "#/definitions/VideoGame"
      operationId: createGame
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetVideoGameResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /videogames/{id}:
    x-swagger-router-controller: game.controller
    
    get:
      description: get a videogmae
      parameters:
        - name: id
          type: string
          in: path
          required: true
      operationId: getGameById
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetVideoGameResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: update a video game
      parameters:
        - name: id
          description: videogame id
          type: string
          in: path
          required: true
        - in: body
          name: videogame
          description: The videogame to update.
          schema:
            $ref: "#/definitions/VideoGame"
      operationId: updateGame
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetVideoGameResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
            
    delete:
      description: delete a videogame
      parameters:
        - name: id
          description: videogame id
          type: string
          in: path
          required: true
      operationId: deleteGame
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /pokemons:
    x-swagger-router-controller: pokemon.controller
    get:
      description: get the pokemon list
      parameters:
        - name: name
          in: query
          type: string
          required: false
        - name: des
          in: query
          type: string
          required: false
        - name: imageLink
          in: query
          type: string
          required: false
        - name: power
          in: query
          type: number
          required: false
        - name: locationLat
          in: query
          type: string
          required: false
        - name: locatonLong
          in: query
          type: string
          required: false
        - name: isCatch
          in: query
          type: boolean
          required: false
        - name: sort
          in: query
          type: string
          required: false
        - name: fields
          in: query  
          type: string
          required: false          
      operationId: getPokemons
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetPokemonListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new pokemon to the list
      parameters:
        - in: body
          name: pokemon
          description: The pokemon to create.
          schema:
            $ref: "#/definitions/Pokemon"
      operationId: createPokemon
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetPokemonResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /pokemons/{id}:
    x-swagger-router-controller: pokemon.controller

    get:
      description: get a pokemon
      parameters:
        - name: id
          type: string
          in: path
          required: true
      operationId: getPokemonById
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetPokemonResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      description: update a pokemon
      parameters:
        - name: id
          description: pokemon id
          type: string
          in: path
          required: true
        - in: body
          name: pokemon
          description: The pokemon to update.
          schema:
            $ref: "#/definitions/Pokemon"
      operationId: updatePokemon
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetPokemonResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: delete a pokemon
      parameters:
        - name: id
          description: pokemon id
          type: string
          in: path
          required: true
      operationId: deletePokemon
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
  /swagger:
    x-swagger-pipe: swagger_raw
    
# complex objects have schema definitions
definitions:

  VideoGame:
    type: object
    properties:
      name:
        type: string
        description: Name of the game
      developer:
        type: string
        description: Developer/Publisher of the game
      gamesystem:
        type: string
        description: Name of the game system
      genre:
        type: string
        description: Name of the genre/style of the game
      year:
        type: number
        description: Release year
      image:
        type: string
        format: byte
    required:
      - name
      - developer
      - gamesystem
      - genre
      - year

  Pokemon:
    type: object
    properties:
      name:
        type: string
        description: Name of the pokemon
      des:
        type: string
        description: Description of the pokemon
      imageLink:
        type: string
        description: Link of the image pokemon
      power:
        type: number
        description: Power of the pokemon
      locationLat:
        type: string
        description: Latitude of the pokemon
      locationLong:
        type: string
        description: Longitude of the pokemon
      isCatch:
        type: boolean
        description: Pokemon is catch or not
    required:
      - name
      - des
      - imageLink
      - power
      - locationLong
      - locationLat
      - isCatch
      
  GetVideoGameResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      name:
        type: string
        description: Name of the game
      developer:
        type: string
        description: Developer of the game
      gamesystem:
        type: string
        description: Name of the game system
      genre:
        type: string
        description: Name of the genre of the game
      year:
        type: number
        description: Release year
      image:
        type: string
        format: byte
    required:
      - id
      - name
      - developer
      - gamesystem
      - genre
      - year
      
  GetPokemonResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      name:
        type: string
        description: Name of the pokemon
      des:
        type: string
        description: Description of the pokemon
      imageLink:
        type: string
        description: URL of the pokemon image
      power:
        type: number
        description: Power of the pokemon
      locationLat:
        type: string
        description: Latitude of the pokemon
      locationLong:
        type: string
        description: Longitude of the pokemon
      isCatch:
        type: boolean
        description: Pokemon is catch or not
    required:
      - name
      - des
      - imageLink
      - power
      - locationLat
      - locationLong
      - isCatch
      
  GetVideoGameListResponse:
    required:
      - videogames
    properties:
      videogames:
        type: array
        items:
          $ref: "#/definitions/GetVideoGameResponse" 
          
  GetPokemonListResponse:
    required:
      - pokemons
    properties:
      pokemons:
        type: array
        items:
          $ref: "#/definitions/GetPokemonResponse"
  
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: Returns 1 if successful
      description:
        type: string
        description: A short comment
    required:
      - success
      - description
  
  GameSystem:
    type: object
    properties:
      name:
        type: string
        description: Name of the game system
      description:
        type: string
        description: Description of the Game System
      image:
        type: string
        format: byte
    required:
      - name
      - description
    
  GetGameSystemResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      name:
        type: string
        description: Name of the Game System
      description:
        type: string
        description: Description of the Game System
      image:
        type: string
        format: byte
      
  GetGameSystemListResponse:
    required: 
      - gamesystems
    properties: 
      gamesystems:
        type: array
        items:
          $ref: "#/definitions/GetGameSystemResponse"
          
  ErrorResponse:
    required:
      - message
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string